api/requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.0
psycopg2-binary>=2.9.9
requests>=2.31.0
pydantic>=2.5.0
